调试

+++ System/Timer.c
@@ -6,7 +6,7 @@
 #include "Serial.h"
 #include "Control.h"
 extern T_angle P_angle;//存储角度测量值
-extern uint16_t P_height;//存储测量高度值
+extern float P_height;//存储测量高度值
 extern float t_yaw;
 extern float t_pitch;
 extern float t_roll;
